airport-service/src/main/java/com/mikhailkarpov/airports/api/AirportController.java
package com.mikhailkarpov.airports.api;

import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.server.ResponseStatusException;

import java.util.List;

@RestController
@RequestMapping("/api/v1/airports")
public class AirportController {

    private final AirportService airportService;

    public AirportController(AirportService airportService) {
        this.airportService = airportService;
    }

    @GetMapping
    public List<Airport> listAirports() {

        return airportService.listAirports();
    }

    @GetMapping("/{code}")
    public Airport getAirportByCode(@PathVariable String code) {

        return airportService.findAirportByCode(code).orElseThrow(
                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Airport with code " + code + " not found")
        );
    }
}
